Move missing-bracket fn to utilities module
